backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
aiofiles>=24.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# RAG and Vector Database
chromadb>=0.4.24
langchain>=0.1.0
langchain-openai>=0.0.1
pypdf>=4.0.0
google-generativeai>=0.3.0
